Make BSP 1st class citizen
