+++ bsp/test/resources/gen-idea-hello-world/build.sc
