+++ bsp/test/src/BspInstallTests.scala
@@ -1,4 +1,4 @@
-package mill.contrib.bsp
+package mill.bsp
 
 import mill.util.ScriptTestSuite
 import os._
@@ -6,14 +6,14 @@ import utest._
 
 object BspInstallTests extends ScriptTestSuite(false) {
   override def workspaceSlug: String = "gen-idea-hello-world"
-  override def scriptSourcePath: Path = os.pwd / "contrib" / "bsp" / "test" / "resources" / workspaceSlug
+  override def scriptSourcePath: Path = os.pwd / "bsp" / "test" / "resources" / workspaceSlug
 
   def tests: Tests = Tests {
     "BSP install" - {
       val workspacePath = initWorkspace()
-      eval("mill.contrib.BSP/install")
+      eval("mill.bsp.BSP/install")
 
-      assert(exists(workspacePath / ".bsp" /"mill.json"))
+      assert(exists(workspacePath / ".bsp" / "mill.json"))
     }
   }
 }
