+++ build.sc
@@ -69,8 +69,7 @@ object Deps {
   def scalaReflect(scalaVersion: String) = ivy"org.scala-lang:scala-reflect:${scalaVersion}"
   def scalacScoveragePlugin = ivy"org.scoverage::scalac-scoverage-plugin:1.4.1"
   val sourcecode = ivy"com.lihaoyi::sourcecode:0.2.1"
-  val ujsonCirce = ivy"com.lihaoyi::ujson-circe:1.2.0"
-  val upickle = ivy"com.lihaoyi::upickle:1.2.0"
+  val upickle = ivy"com.lihaoyi::upickle:1.2.1"
   val utest = ivy"com.lihaoyi::utest:0.7.4"
   val zinc = ivy"org.scala-sbt::zinc:1.4.0-M1"
   val bsp = ivy"ch.epfl.scala:bsp4j:2.0.0-M4"
@@ -480,16 +479,6 @@ object contrib extends MillModule {
     def testArgs = T(scalanativelib.testArgs())
   }
 
-  object bsp extends MillModule {
-
-    override def compileModuleDeps = Seq(scalalib, scalajslib, main, scalanativelib)
-    def ivyDeps = Agg(
-      Deps.bsp,
-      Deps.ujsonCirce,
-      Deps.sbtTestInterface
-    )
-  }
-
   object artifactory extends MillModule {
     override def compileModuleDeps = Seq(scalalib)
   }
@@ -546,6 +535,14 @@ object scalanativelib extends MillModule {
         )
     }
   }
+}
+
+object bsp extends MillModule {
+  override def compileModuleDeps = Seq(scalalib, scalajslib, scalanativelib)
+  def ivyDeps = Agg(
+    Deps.bsp,
+    Deps.sbtTestInterface
+  )
 }
 
 def testRepos = T{
@@ -677,7 +674,7 @@ def launcherScript(shellJvmArgs: Seq[String],
 }
 
 object dev extends MillModule{
-  def moduleDeps = Seq(scalalib, scalajslib, scalanativelib)
+  def moduleDeps = Seq(scalalib, scalajslib, scalanativelib, bsp)
 
 
   def forkArgs =
